feat(vm): 实现减法、乘法和除法操作并优化数值转换
- 在 call 操作中添加对 sub、mul 和 div 方法的支持，支持整数和浮点运算
- 修复 i32 和 i64 到 f32 的转换操作，直接使用 Value::f32 避免不必要的 f64 中间转换

+++ projects/nyar-vm/src/vm/ops/i32.rs
@@ -189,16 +189,16 @@ impl NyarVM {
     #[inline(always)]
     pub fn execute_i32_to_f32_s(&mut self) -> Result<Option<usize>, NyarError> {
         let v = self.pop()?;
-        let r = (v.as_int() as i32) as f32 as f64;
-        self.push(Value::float(r))?;
+        let r = (v.as_int() as i32) as f32;
+        self.push(Value::f32(r))?;
         Ok(None)
     }
 
     #[inline(always)]
     pub fn execute_i32_to_f32_u(&mut self) -> Result<Option<usize>, NyarError> {
         let v = self.pop()?;
-        let r = (v.as_int() as u32) as f32 as f64;
-        self.push(Value::float(r))?;
+        let r = (v.as_int() as u32) as f32;
+        self.push(Value::f32(r))?;
         Ok(None)
     }
